requirements.txt
# Core dependencies
strands-agents
strands-agents-tools
strands-agents[anthropic]
strands-agents[openai]
strands-agents[writer]
strands-agents[ollama]
python-dotenv



# Optional - uncomment as needed
# strands-agents[otel]  # For observability/tracing
# duckduckgo-search     # For web search tool
# requests              # For HTTP requests
# boto3                 # For AWS services